// src/app/constants/route.enum.ts
export enum Route {
    EMPTY = '',
    LOGIN = 'login',
    SIGNUP = 'signup',
    HOME = 'home',
    TEACHERSU = 'teachersu',
    STUDENTSU = 'studentsu',
    PROFILE = 'profile',
    PROGRESS = 'progress',
    SETTINGS = 'settings',
    MATHEMATICS = 'mathematics',
    READINGWRITING = 'reading-writing',
    COMMUNICATION = 'communication',
    TEACHERPROFILE = 'teacherprofile',
    PROGRESSMONITOR = 'progressmonitor',
}